Initial migration of the typescript code from transition-legacy
This code was in a private repository, where other applications were
developped as separate branches or directories. This commit contains the
part of that other repository to open source with the Transition
application.

The code from the legacy repo was contributed by the following
contributors, ordered by total number of commits.

Co-authored-by: Pierre-Léo Bourbonnais <leo.bourbonnais@polymtl.ca>
Co-authored-by: Geneviève Bastien <gbastien@versatic.net>
Co-authored-by: Vincent Chabin <vincent.chabin@polymtl.ca>