demo_survey: Let the demo_survey be the start of the client app
Add a package.json file to the demo_survey example application and add
the workspace in the main package.json.

Update evolution-legacy's app-survey.js file to return a function the
callers can call to bootstrap the application after configuring the app.

Add a .babelrc file to demo_survey and compile the sources with babel

Add a app-survey.js file in demo_survey to configure and start the
application.

Move the webpack.config.js file to the demo_survey package to build the
client app from this package.