Make the InterviewResponses type generic
In order for survey projects to be type checked, the responses must be
exhaustively typed, with each field and its value type. A response set
has a household with persons and home fields by default, but each of
those types, as well as the root response type can be customized for a
survey. So the InterviewResponse now takes 4 types: CustomSurvey,
CustomHousehold, CustomPerson and CustomHome.

Making this type generic has a lot of side effect on the interview
attributes, WidgetConfig and widgets themselves, who now should be
generic too.

The WidgetConfig's string path can now be typed to match only the
elements of the response type.

The demo_survey project is updated so the typed widgets work, with
the custom typed correctly defined.