Make the validations type generic
The validations should have the same fields as the responses, except as
boolean. Create a generic type that is a function of the custom types of
the responses, with boolean values.

Changing this type required to cast the widgets used by the application
to `any` as generic types, otherwise it gives compilation error because
of the withTranslation HOC used by the component, which does not seem to
correctly propagate the generic types. But since the widgets are only
visual and do not require direct access to the generic fields, it is a
small price to pay to be able to type check the validations and fields
where it really matters.