survey: Let interviews be associate with participant instead of user
Fixes #78

* Fill the participant table with the user's data corresponding to each
current survey and switch the user_id column to participant_id column.

* Update interview types to have a participant_id field instead of
  user_id

* Add a participant authorization middleware to verify that a
  participant is only allowed access to his own interview. The previous
authorization middleware is renamed to userAuthorization and the default
role now has no access as a user does not have an interview anymore.

* Copy the server and client app setup files in root /src folder to
  /src/apps/participant and edit to user the participant authentication
model and participant authorization middleware. Routes setup file is
also moved to this app folder and only includes the survey routes for
the participant application.

* Update all tests