Add an admin application setup and demo_survey app
Completes #51

The admin application setup files are in `apps/admin` client and server
folders in evolution-legacy.

There's a new AdminHomePage, which displays the possible links to pages
for a user.

In the demo_survey project, the application startup scripts are in the
admin/ folder.

There are new yarn scripts for the admin app `yarn start:admin` for the
server, `yarn build:admin:dev` and `yarn build:admin:prod` for the
client. To have apps run on separate ports, use `yarn start --port 8082`
to specify the port, for either the participant or admin app.

** BREAKING **

The previous commit removed the administrative functionalities. With
this commit, it is brought back, but as a separate application. To have
this application running, a project needs to

* add a bootstrap file for the administrative client and the administrative
 server
* Add new yarn scripts to run the server and build the client
* Add a webpack config file for the admin client, pointed to by the
  admin yarn script.

See the demo_survey app for an example to setup the server app.