db: Fix set participant migration
Do not use the participantsDbQueries directly in the migration, as it
does not use the same knex object. It caused a deadlock where the query
was waiting for the current transaction to finish, which does not end as
the it waits for the user to be inserted.

Applications without any interviews yet will not be affected. Others
will not have been able to run the migration, so it fixes the issue for
those.