Update Transition, with i18next package upgrades
To match Transition's package versions, we upgrade the i18next package
versions as follows:

* i18next: from 20.2.2 or 19.5.1 to 22.4.15
* i18next-browser-languagedetector: from 5.0.0 to 7.0.1
* react-i18next: from 11.7.0 to 12.2.2

With the i18next major release to 22.0, typescript types have been much
updated. The main change is that TFunction can now optionally return
null, and it allows null by default. Since most of our usages of t
functions are expected to be `string | undefined`, we can set the
`returnNull` option to `false`, but for compilation to work correctly,
it needs to be typed. As suggested in the i18next documentation, we add
a i18next.d.ts type file with the returnNull set to false.

Also update the `.vscode/settings.json` file with i18n ally
configurations.