i18n: Allow the translatable strings to be functions taking `t` argument
Any translatable string in the WidgetConfig's types, can now be either a
string, an object, with the key being the language and value a string or
function. We add the possibility that it might be a function that takes
the `t` function as parameter, as well as the interview, path, user.
This allows labels and other translatable strings to access the complete
range of the i18next features. The new type for those strings is
I18nData<CustomSurvey, CustomHousehold, CustomHome, CustomPerson>.

Add a `translateString` function to the helper that takes a
translatable string value, a t function, the interview, path and user as
parameters and returns the proper value. This function can be used to
process the labels.

These new types are not being used yet.