widgets operations: Check that path is a string instead of widget type
Fixes #99

Any widget type may have a path that can be used in conditional or labels,
etc. Add an optional `path` field for widget types that didn't have one.

Check that the path is of type string before interpolating it instead of
only for question and group types.

Add unit tests for a widget of type text with a path.