inputs: Support html in choices label
fixes #95

fixing inputRadio and inputCheckbox to be able to render html in the label of each choice
We are going to implement the dangerouslySetHtml without the containsHtml attribut.
Our major concern was about a potential security breach. We were worried about a malicious input from users.

That's why we used the dompurify library which prevent any XSS attacks like described on this page : https://victoria2666.medium.com/the-implications-of-dangerouslysetinnerhtml-in-react-64e2d02af66a

Moreover, if at one point in the futur, the field __html contained in dangerouslySetHtml is subject to users' input (meaning that if the variable passed to __html is the result of a user input from the frontend), it will be important to limit and regulate the input field to ensure the program safety.