question: Extract the question div to a widget
The label, error message, help popup link and input itself are extracted
to a separate widget, responsible for the layout of those elements.

The widget collapse is the responsibility of this new widget. When the
label is clicked and the widget can be collapsed, it toggles the
collapse state. This behavior previously did not work.