question: Add unit tests for widget accessibility (a11y)
Add dependencies to `jest-axe`, which is a unit test library that works
with the React Testing Library to make sure widgets have no violations
to the WebAIM best practices.

The Question widget is tested with this framework, for each widget type.
This will ensure there are no regression to the accessibility. For now,
though, InputRange and InputMultiselect are not accessible and this test
is disabled for them.