slider: Make the InputRange widget accessible
fixes #38

The InputRange uses react-input-range package, which create a component
without an ID. Other react widget libraries, such as rc-slider, who
could eventually replace react-input-range also do not have id. This
results in an orphaned label and another unlabeled inputs.

For this input type, we use the ariaLabelledBy prop in the widget itself
and the label is replaced by a `div` whose ID is the ID of the widget,
suffixed with `_label`.