multiselect: Make the InputMultiselect widget accessible
The Multiselect uses the react-select widget, which does not assign the
ID to the input and thus cannot use the label for id. It causes an
orphaned label and an unlabeled input. Similar to the InputRange, we put
the label in a `div` with an id suffixed with `_label` and use the
`aria-labelledby` prop to specify the label div.