inputSelect: Add line separator for groups without label
Fixes #130

Select choices can be separated in groups with group labels. If the
label is set to an empty string, instead of adding an indented optgroup
with an empty label, add a disabled horizontal line option to separate
the group from the rest.

Inpired from this stackoverflow answer: https://stackoverflow.com/questions/899148/html-select-option-separator

Update the car driver's parking question to split in group with empty
label. And re-activate the dwelling type question, for an example with
groups with labels.