logging: Add a table and queries to log user accesses
The table logs the user_id, the interview_id, and the mode (validation
or edit). It also has a field to count the number of update queries and
dates at which the interview was first accessed and last updated.

The db queries and tests are added to insert/update records when an
interview is opened/updated.