logging: add a logging middleware when opening/updating an interview
fixes #43

Add a queryLoggingMiddleware file, with middlewares that can be used
either in validation or edit modes and log to the database any access by
the current user. These middlewares are used in the express routes when
the interview is first opened and when it is updated.

The admin routes use the logAccessesMiddleware and the participant
routes use the default middlewares, which do nothing.