homepage: add a consent checkbox to agree to usage of data
Fixes #128

Create a new widget with a consent checkbox and the start button.

By default, the main:AgreementText in evolution is empty, so there won't
be a checkbox. If the `survey:homepage:AgreementText` translation is
defined in a project, it will be displayed next to a checkbox. This
checkbox must be checkbox for the button to go to the main login page.

If the checkbox is not checked, the `survey:homepage:ErrorNotAgreed`
text is displayed. It falls back to a default error message indicating
to check the box.