Only update map bounds if the last action is a shortcut selection
Previously, if we clicked on the map, the map would zoom and pan to this location. If we clicked from really far away (i.e. imprecise click), the map would zoom close to the marker, yet it would say that it's not precise enough. Now, we make sure that the action is shortcut before doing this.