InputMapFindPlace: display a warning in red if the geocoding query is too imprecise
If there is only a single result, we check if it's precise enough using the location type information returned by the geocoding API. The Google Maps geocoding API returns an array of location types for each result. For a result, if all of its location types are deemed too imprecise, then this result is considered to be imprecise and an error is shown to the user.