Input Choices: add alignment options
Add more options to control how and in which order the options for radio and checkbox type questions appear.

Input radio and checkboxes can specify a number of columns or rows and
whether to align the choices vertically or horizontally. It is also
possible to specify custom alignment, to group elements in columns or
rows.