Remove 'Loading...' label for button, instead just disable the button
Ideally this should be done in InputButton, but some surveys still use the Button.js class from evolution-legacy.

When a survey was loading, this button used to be changed to a label (ex: "Loading..."). However, since this label was smaller than the button, the content below would be moved slightly which would be quite annoying.

Now, instead of replacing the button with a label, we simply disable it so it is not clickable anymore. Since the height does not change, the flickering problem is solved.

Ideally, the property would be renamed disabledWhenRefreshing, but since this class will eventually be replaced by InputButton, I am keeping it as-is to avoid breaking changes.