Add an evolution-interviewer node module for interviewer support
Fixes #175

This interviewer module adds roles for interviewers, which are users allowed
to read or update any interviews and interviewer supervisor, which have the
same permissions as interviewers, but also manage interviewers.

This module also adds support for an interviewer mode in the interview,
which can be used to change labels for this mode, or by widgets.

To use, the server admin application needs to call the `roleDefinition`
default function to add the roles.

The frontend admin application needs to import the `interviewerSupport` file
and add the interviewer options to the application configuration.