helpers: Add a function to replace shortcuts by the original place
Add a helper function which takes the interview and a path to a place.
For all places using this place as a shortcut, update their
names/geography/shortcut as follows:

* The first other visited place is updated to take the values from the
  original visited place
* All other visited places are updated so that the shortcut now points
  to this first other visited place.

This function returns the valuesByPath to update and the paths to unset
(the shortcut of the first other visited place, if necessary).

The deleteVisitedPlace helper function now calls this helper to avoid
the place being deleted without updated the shortcuts first.