interviewer: Add interviewer monitoring route and component
* Add a frontend monitoring component, which allows to select a start
  and end date, then update the interviewer cache and downloads the file

* Add a db query to fetch all interviews touched by interviewers, within
  a certain period and as a paginated list

* Application that use this interviewer monitoring module need to call a
  function in `interviewer.routes.ts`, which will put in place the
necesary routes for interviewer data. This function receives as
parameter an object with the following functions:

{
  `data: (responses) => object` which receives the responses of an
individual interview and should return an object with specific fields to
cache for this interview.

  `aggregate: (dataForInterview: any[]) => object` which receives all
the parsed interview data for a single interviewer and can aggregate
statistical data from those. For example, if a field for the interview
is a counter (number of x), the aggregate function could return a field
that would be a sum, or an average of those counters.
}