add a way to keep the assigned value intact
when toggling between hide/show conditional status, this is useful if you want to use two different widget questions for the same value, and toggle between them according to a condition. Without this config, whenever the condition is switched back to show after a hide, the hide assigned value is replaced by the default or empty value. This is also useful to hide a question when a condition is met, but keep its assigned value active between hide and show toggles