survey: Add the submitted parameter after updating the survey
The submitted parameter was only set when the `{ _all: true }` was the
only updated value and no server call was necessary. But if there are
validation updates to do, like after removing the value for a field with
side effects, a call to the server is necessary and the next
rendering does not scroll to the invalid questions, if any. If the user
action was a submit, the update after the server call also sets the
submitted to `true`.