validations: Add warning if responses were modified since copy
fixes #240

The validateInterview GET query returns a `validationDataDirty` boolean
flag that is `true` if the validated_data's `_validatedDataCopiedAt`
field is lower than the responses' `updatedAt` field, or not set (for
previous validations).

The message is displayed on top of the interview summary and invites the
validator to reset the interview data.