admin: Support materialized views for monitoring
fixes #296

This adds support for materialized views in surveys. The server can
register the view with the `registerView` method, with the corresponding
SQL query at startup. If the view already exists, if verifies if the
query is the same as previously. If not, the view will be dropped and
re-created.

The view can be refreshed, queried with all or a subset of the fields,
or using a groupBy/count query.

This commit just provides the API for view support. Surveys can then
create their own views and query them. It will provide a less memory
intensive way to parse all surveys for various information for
monitoring purposes.