db: Add an index on the responses.accessCode field
fixes #315

This increases performance of listing interviews by more than 10 times.
From local benchmarks in a 15K record database, it does not affect
perceptibly the performances of the responses updates or the survey
flow. The database update/insert times are within 2% with/without the
index, less than the normal variability.