migrations: Remove duplicate interviews before setting unique contraint
The migration which sets the survey_id and unique constraint is updated
to remove duplicate interviews for a same participant. The interviews
are deleted and only the one with the longest response field is kept.

The migration itself is modified. It would have been impossible to run
this migration if there was duplicate interviews anyway. For instances
that have already successfully run it, it has no effect.