generator: Generate the labels from a separate sheet
Copied from the `fillLocales.py` script. The sheet contains a `key` column,
a 'md' column to keep markdown or convert to html.
The rest of the columns should be in the [lang][_[context]] format,
where `lang` is the language and the optional `_context` is a suffix
that can be appended as a context to the label, for example `_one` to
differentiate between a 1 or multiple respondents interview.