generator: Generator widget labels with person data
fixes #432

This requires 2 optional columns in the Widgets sheet:
`hasPersonsCntLabel` and `hasGenderedSuffixLabel`.

Putting any of those to `True` will add the proper context values to the
label's `t` function call, with either genderedSuffix, countPerson, and
nickname.

For now, this feature works only for the sections after household, when
all the persons in the interview have been defined. The texts can either
be defined in the "Widgets" sheet under the "fr" or "en" column or in
the "Labels" sheet.