backend: Let serverFieldUpdate return a redirection URL
The server field update callback can now also return an array with the
fields to update, as well as a string representing an URL to redirect
to. This URL is returned from the `updateInterview` call.

The `/survey/updateInterview` route can return a response with the
status of `redirect`, along with a `redirectUrl` field. If so, the
client should redirect to this URL directly.