Use an http response code handler for error and unauthorized
fixes #53

When the page is unauthorized, verify the user authentication to make
sure he is properly logged out of the application to avoid looping to
the unauthorized page. It also refreshed anything that might have
changed server side concerning the current user.

Use a specific `UnauthorizedPage` for evolution, to allow a layout that
fits the rest of the application.

On the admin side, the `InterviewListComponent` also handles http
statuses.