playwright: Correctly identify each range widget
Add a sliderColor parameter to the inputRangeTest to identify the
slider horizontal rule by its class, that depends on the color.

Use the inner the parent container as the base locator to identify the
right slider object, as they are not identified by ID, but only by
class and some may have the same values/label.