playwright: Move the survey ID object detection and replacement to its own file
Add a `SurveyObjectDetector.ts` file which contains the code to detect
the functions to detect the survey objects from the update data.

Add code to detect the journey IDs, visited places IDs, activePerson,
activeJourney and activeVisitedPlace.

This file also contains the logic to handle those values and replace
them in the paths.