playwright: Make sure widgets to test are in the viewport
When taking a screenshot on failure, only the viewport is taken and we
should make sure the element is visible in the current viewport.

To make sure that the focusOut does not scroll to the top, we click on
the extreme left of the screen instead of a header element.