playwright: Add context test parameters to allow multiple parallel runs
Move the survey object detector functions to a class, that can be
instantiated independently for each test, to make sure each parallel run
has its own survey objects.

Move the page, widgetTestCounters and objectDetector to a
`CommonTestParameters` type with a `context` field that will be passed in
parameters to all tests. This ensures that tests do not share these
values and thus conflict their respective pages.

Add a code snippet in the README to show how to write a simple basic
test.

Fix the demo_survey tests to use the test parameters