Drop the generic types from the Interviews types
fixes #530

The initial intention was to be able to type each specific survey
responses, but we will rather refactor Evolution to have common objects,
in addition to custom fields, so the generic are not necessary anymore.

For now, where necessary, the cast is forced to any so it can compile,
but as we move to use the surveyObject types, we can update those.