widgets: Move the Button component to evolution-frontend
Update references and add tests

** Breaking change **

To avoid the actions and callbacks of the button to have undocumentable
and possibly arbitrary `this` object, those widget configs do not bind
the `this` anymore. Thus, they need to receive all the necessary
parameters that used to be assumed in the props.

So, for widgets of `type: 'button'`,

1. The parameters of the `action` function are changed from `(section, sections,
   saveCallback)` to `(callbacks, interview, path, section, sections,
   saveCallback)`

2. The parameters of the `saveCallback` function are changed from `()`
   to `(callbacks, interview, path)`.

Where the `callbacks` parameter contains the redux actions that may be
called by the function. See the `ButtonWidgetConfig` type in the
`WidgetConfig.ts` file for details.

**When surveys will update evolution, they will need to update their
button `action` and `saveCallback` function.**