generator: Validate that row column exists before assigning
In python `row['something']` does not return None if 'something' does
not exist in row, it throws an exception, so we need to make sure the
fields are available.

Recently added columns caused problems with previous excel files, so the
checks are added to those here.

TODO We should validate this for each column, not just the new ones.