inputRange: add a `includeNonApplicable` property
fixes #539

To avoid the participant having to select a position for a question that
does not apply to him, there should be a "Non applicable" option that
the user can select that will disable the input range. The answer in
this case, instead of a `number` will be a `string` with value `na`.

Also include the generation of this property from the generator, with
the `includeNotApplicable` boolean column.