widgets: The type of a widget config cannot be a function
If a widget can have multiple types, it is preferable to use 2 widgets
with conditionals mutually exclusive, or create a specific widget that
can be multiple types (eg. radio or select), but allowing any type in a
single widget can be tricky as there isn't much in common between a text
and a radio widgets, for example, it requires to put all the possible
configs for all types.