exports: Allow to export validated or original responses by object
fixes #557

The export task receives an option parameter, which has a `responseType`
field, with values 'validatedIfAvailable' or 'participant'. The exported
file names will be prefixed with either validated or participant
depending on the response type.

NOTE: The routes and widgets are in the individual surveys in this
branch, they will need to be updated.