server: Add a deferred server update callback mechanism
fixes #582

Add an extra parameter for server update callbacks, so that callbacks
that have time-consuming operations (like routing calculations) can
register the operation, which can then execute without blocking the
current client update call.

The callbacks can return quickly empty or other values, but register
an operation, identified by a `opName` and `opUniqueId`. The `opName`
ensures that only one operation for that name will be considered. The
`opUniqueId` allows to cancel any previous operation with the same name.
The `operation` to register is a function that receives a `isCancelled`
function and returns a promise that will run in the background. The
promise should return the values by path to set in the interview. The
interview will be saved with the new values and those can be sent back
to the client at the next update call.

The deferred update callback will save the updated values in the user's
session and they will be added to the returned server values in the next
client response.

Example usage of this mechanism

```
field: 'accessCode',
callback: (interview, value, path, registerUpdateOperation) => {
    // Use the update operation to set the city to Calgary, but 5 seconds later
    registerUpdateOperation({
        opName: 'city',
        opUniqueId: 1,
        operation: () => new Promise((resolve) => {
            setTimeout(() => {
                resolve({
                    'home.city': 'Calgary'
                });
            }, 5000);
        })
    });
}
```