sections: Make sure the invalid widget has an ID before focusing
Some widgets, like InputRange, do not have an ID (and cannot have one, as
the library used does not allow to set it), but they can be invalid. To
automatically scroll to the invalid question, we make sure the input has
an ID before focusing on it. Otherwise, we use the other code path that
just scrolls the page to make the widget visible.