playwright: fix `inputMapFindPlaceTest` when there is only 1 choice
fixes #568

The test can pass an extra parameter: `expectMultiple` which, if `true`
selects the first returned location from the select and confirms the
location (previous behavior).

If `false`, it simply waits for the widget to have the 'question-valid'
class and checks that the text to identify the location is correctly shown.