playwright: Allow `waitTextVisible` to also validation invisibility
An extra parameter to this test, `isVisible`, allows to specify if the
text is expected to be displayed or not (defaults to `true).

Careful though, to call this test only once previous tests have
validated that the expected page is shown, otherwise, the text may just
not be visible yet, but may appear later.