map: Use the user's current language when loading the map
fixes #593

This uses the user's current language for the map language instead of
the defaultLocale. It falls back to the default locale if not set.

The map configuration must be fixed throughout the survey, as we cannot
call the `useJsApiLoader` a second time with different parameters. The
configuration is global and determined only when the first map is shown.
After that, changing the language will not change the language of the
map, but if the page is refreshed (with F5), it will be updated.

see https://stackoverflow.com/questions/7065420/how-can-i-change-the-language-of-google-maps-on-the-run