survey: Update the interview responses types and add journeys
The responses type, during the survey execution, are updated to use the
corresponding main object's attributes type, without the composed type
which are objects instead of arrays in the responses json object.

** Breaking change **

This also adds the journeys object under the Person, under which the
visited places and trips are now located. This will break all previous
surveys which put the visited places directly under Person. These
surveys, if they used the helper functions from Evolution, will need to be
udpated to add the journeys in their path.

Surveys that implement their own helpers and don't type-check should
still work correctly.

The demo_survey project is updated to show what may need to be done to
achieve the journey integration in a survey project.