segments: Add a widget for the modePre question of the segment section
This adds a generic segment mode pre question, that generates a radio
question widget offering a choice of mode category for a given segment.

All labels are fully localizable, with english and french translation
provided.

The carDriver option is only shown if the person has a driving license
or is old enough to have a driving license if the information is not
available. If age is not available, it is shown.

Add yml translation files for survey (generic survey translations) and
segment (specific for the segment section)