survey: Add an ErrorBoundary component around survey
fixes #466

Add a SurveyWithErrorBoundary component that wraps the Survey component
in an ErrorBoundary component which catches any uncaught exception and
redirects to the SurveyErrorPage.

This new component replaces the Survey in the router.

Add a `onRedirect` prop to the SurveyErrorPage component which allows to
reset the error boundary when the link is clicked.