config: Type the Evolution-specific configuration options
Add a file in evolution-common, `project.config.ts`, which extends
chaire-lib's project configuration types.

Start documenting a few of the evolution project's configuration
property and set default values for them.

Import this new project.config file instead of the chaire-lib's one, to
be able to call directly the typed properties.