admin: Refactor InterviewSummary to have links even with errors
fixes #664

This adds AdminErrorBoundary around the InterviewSummary, the
InterviewMap and the InterviewStats component to make sure other
components can be displayed even if there is an error in one of them.

Also the InterviewSummary is now responsible to display the
ValidationLinks so that those links are available even if the summary
page is not set for the interview. This means we can invalidate
interviews and edit them in validation mode.