mapFindPlace: Allow multiple geocoding query string and zooms
fixes #665

The `geocodingQueryString` function now returns either a string or an
array of objects containing a `queryString` and a `zoom`. The first
object is geocoded and if results are returned, they are used and the
map zooms at the specified level. If no results are returned, the next
query string object is geolocated, etc.