db: Add geojson filter to interview's `getList`
The filter value type can now be a geojson polygon feature. The requested
field is expected to be a geography feature field. The filter will add a
query to return the interviews for which the field is non null and
contained in the requested polygon.

Add unit tests for this feature.