admin: Add validation filter on home geography
fixes #666

The filter on home geography is an Text area which takes a polygon
feature, or a feature collection with a polygon feature. The textarea
cannot really be edited, the feature needs to be copy pasted in the box.
This avoids entering bad data and is simpler for the validation.