widget preparation: Return valuesByPath when no widget and validation
In the validationOnePager section, there may not be widgets defined, so
no widget to prepare, but we need the original valuesByPath from the
request to update the validity or completion of the survey.

A FIXME mentions to look into this code, as `valuesByPath` may be what
we want in all cases, not just for validationOnePager.