serverFieldUpdate: Run the callbacks on validated_data as well
Add a field to the `ServerFieldUpdateCallback` type to indicate whether
to run this callback when the updated value is in the validated_data
instead of responses. Defaults to the current behavior of `false`.

When the survey is updated in validation mode via the admin interface,
the server field update callbacks will be executed only if explicitly
mentioned.