transition: Add a `getTransitSummary` function
fixes #684

This function will retrieve a summary for a transit route, ie the number
of transit routes that can be used for the trip, as well as a list of
lines used.

For now, this function uses the Transition API to call the public
`summary` call.

Add unit tests for the new functions.