adminExports: explode arrays of object in csv exports
fixes #694

When a survey attribute is an array of complex types (object), instead
of simply joining the values with `|`, we explode the paths to display
each of its fields, ie `attribute.0.field1`, `attribute.0.field2`,
`attribute.1.field1`, etc.

For arrays of simple type (strings, numbers, etc), the current behavior
is kept.