admin: Use deck.gl for the InterviewMap component
fixes #668

Add dependencies to `deck.gl` packages and `popmotion` for arrow
animation.

Add the mapDefaultCenter to the project's configuration type.

Change the InterviewMap component to use deck.gl instead. Add an
AnimatedArrowPathLayer to display of path layer with moving arrows
showing the direction of trips.

Add admin styles from the od_nationale_2024 survey for proper layout of
the admin interface.

** From an original commit in a survey by Pierre-Léo Bourbonnais, ported
  to evolution by Geneviève Bastien **

Co-authored-by: Pierre-Léo Bourbonnais <leo.bourbonnais@polymtl.ca>
Co-authored-by: Geneviève Bastien <gbastien@versatic.net>