admin: Add all trips and visited places to the map
Add a function to generate a bezier curve LineString from an array of
points for trips.

Add an admin helper function to retrieve the visited places and trips
from the interview's response data. This function should be deprecated
soon as the validations move toward using objects instead of the json
response object.

Add an application configuration option to get the visited places and
trips for a specified survey.