segments: Add the _isNew property to the Segment type
This property is specified to differentiate between new segments and
those that were already saved and completed once.

Add origin and destination to the `tripId3P2` object, so that the test
case interview has a complete chain of more than one trip. Test that
need no origin/destination can just clone the interview and delete the
origin/destination attributes.