segments: Add helpers specific to the segment section
Add types for loopActivities and simple modes (ie individual/private
modes that are often used for many trips).

Add helpers for survey time to check whether a trip is the return trip
of a simple single chain, ie origin of first trip is same as the
destination of second trip and the mode used is a simple mode. Another
helper returns the unique segment of the previous trip of the chain.

Add unit tests for these functions.