segments: Add a widget for the mode question of the segment section
This adds a generic segment mode question, that generates a radio
question widget offering a choice of the modes, given the selected
category.

Labels and mode selection match the mode values of the segments, labels
are fully localizable, with english and french translation provided.

Some modes have extra conditional function, for example, `wheelchair`
will only be present for persons that may have disabilities.

We add a map where each mode is mapped to a modePre value. When the user
selects a `modePre` value, the modes are displayed. If there's only one
choice available for a given modePre, it is automatically selected.