segments: add a widget for the same mode for return trip
This widget is a boolean radio widget that asks if the mode is the same
as the reverse trip. This will be displayed on new segments, only if the
current and previous trip are part of a simple chain (origin of previous
is same as destination of current) and the mode used is a simple one.

Once the segment is saved, it is not new anymore and this value will be
erased in later updates.